requirements.txt
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.1.1
Shapely==1.7.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
Werkzeug==1.0.1
